remove bragi_features_tes file
