+++ tests/tests.rs
@@ -55,7 +55,6 @@ mod bragi_poi_test;
 mod bragi_stops_test;
 mod bragi_filter_types_test;
 mod bragi_synonyms_test;
-mod bragi_features_test;
 
 use docker_wrapper::*;
 use std::collections::BTreeMap;
@@ -291,5 +290,4 @@ fn all_tests() {
     bragi_stops_test::bragi_stops_test(ElasticSearchWrapper::new(&docker_wrapper));
     bragi_filter_types_test::bragi_filter_types_test(ElasticSearchWrapper::new(&docker_wrapper));
     bragi_synonyms_test::bragi_synonyms_test(ElasticSearchWrapper::new(&docker_wrapper));
-    bragi_features_test::bragi_features_test(ElasticSearchWrapper::new(&docker_wrapper));
 }
